+++ src/componets/MangasDetail.jsx
@@ -3,7 +3,7 @@ export default function MangaDetails() {
     const description=useSelector(store=>store.manga.manga.description)
     return (
         <main>
-            <div className='text-[#424242] text-lg p-8 text-justify '>
+            <div className='text-[#424242] w-[80%] mx-auto text-lg p-8 text-justify '>
              {description}
              </div>
         </main>
